requirements.txt
fastapi
uvicorn
youtube-transcript-api
openai
googletrans==4.0.0-rc1
python-dotenv
google-generativeai